# Add specialized methods for determinant of small matrices