We currently have specialized methods for 2x2 and 3x3 matrices when it comes to matrix multiplication. This is because the overhead in calling BLAS is significant in these cases. There is currently no such optimization for determinants. This PR implements specialized methods for 2x2, 3x3, 4x4 matrices. 

The motivation for this is that I noticed in my finite element code that calculating the determinant of the Jacobian matrix took a significant time in the whole analysis.

The following benchmark shows the difference between master and PR:

```julia
function f(N, n)
    A = rand(n,n)
    @time for i = 1:N
        det(A)
    end
end
```

### Master
```julia
f(10^5, 2);
# 0.742423 seconds (3.10 M allocations: 521.851 MB, 9.48% gc time)
f(10^5, 3);
# 0.756738 seconds (3.10 M allocations: 529.480 MB, 8.73% gc time)
f(10^5, 4);
# 0.756825 seconds (3.10 M allocations: 534.058 MB, 8.77% gc time
```

### PR
```julia
f(10^5, 2);
# 0.001552 seconds
f(10^5, 3);
# 0.001898 seconds
f(10^5, 4);
# 0.003764 seconds
```

Does anyone know if these "naive" formulas are much more sensitive to catastrophic cancellation than the ones from the factorization?