@andreasnoack 2015-08-04
Looks good to me, but please add some tests?

It appears that the Mac build is dead right now.