@KristofferC 2015-08-04
Ok, I will add tests