@KristofferC 2015-08-04
Added a test. Is it reasonable or should there be more stuff there?