@andreasnoack 2015-08-04
I think the tests look fine, but similarly to the comment in your `inv` pr, I think the element type should be restricted a bit.