@IainNZ 2015-08-04
```julia

max_error = 0.0
max_error_A = zeros(4,4)
for s in 1:1_000_000
    A = rand(4,4)
    detA = det(A)
    newA = det4x4(A)
    err = abs(newA - detA)
    if err >= max_error
        max_error = err
        max_error_A = A
    end
end
@show max_error
println(max_error_A)
```

```
max_error = 7.771561172376096e-16
[0.9769920685752007 0.26462435683121344 0.7660953256016576 0.8408033726879165
 0.09664687603968702 0.9475475924467718 0.8423010563873028 0.3578097149120094
 0.7329460495767071 0.9920844560834485 0.056002919177041255 0.900890947141922
 0.8009609939914129 0.7490580692712308 0.15617418215311396 0.24696011176782107]
```

```
julia> 3*eps(Float64)
6.661338147750939e-16

julia> 4*eps(Float64)
8.881784197001252e-16
```

Seems pretty good for stability