@KristofferC 2015-08-04
Hmm, it seems there is some problems with `det` for LU factorizations. This line https://github.com/JuliaLang/julia/blob/master/base/linalg/lu.jl#L155 is very slow. Something with the broadcasting that makes it slow.

I rewrote it a bit which makes it a lot faster for small matrices.:

```julia
using Base.LinAlg

function det_new{T,S}(A::LU{T,S})
    n = Base.LinAlg.chksquare(A)
    A.info > 0 && return zero(typeof(A.factors[1]))
    P = 1.0
    c = 0
    @inbounds for i = 1:n
        P *= A.factors[i,i]
        if A.ipiv[i] != i
                c += 1
        end
    end
    sig = (isodd(c) ? -one(T) : one(T))
    return P * sig
end

function f(N, n)
    A = rand(n,n)
    F = lufact(A)
    @time for i = 1:N
        det(F)
    end
    @time for i = 1:N
        det_new(F)
    end
end

f(10^5, 4)
# 0.676559 seconds (2.50 M allocations: 491.333 MB, 5.61% gc time)
# 0.010234 seconds (300.00 k allocations: 13.733 MB, 19.00% gc time)

f(10^5, 5)
# 0.676559 seconds (2.50 M allocations: 491.333 MB, 5.61% gc time)
# 0.010234 seconds (300.00 k allocations: 13.733 MB, 19.00% gc time)