@KristofferC 2015-08-04
I also noticed that `chksquare` is type unstable for LU factorizations. Changing the first line in `det_new` above to `n = size(A, 2)` gives

```julia
f(10^5, 4)
  0.624297 seconds (2.50 M allocations: 491.333 MB, 5.07% gc time)
  0.000645 seconds
```

which is a 1000x speedup.