@KristofferC 2015-08-04
Doing my change to `det` for lufact means that the handcoded determinants in this PR now are only ~10x faster instead of 200x.