@ivarne 2015-08-04
Off topic Github trick:

@KristofferC if you hit the `Y` key before copying links to specific code lines, you get the full link with SHA, so that other changes to `base/array.jl` doesn't make the reference point to other code than what you intended.