@KristofferC 2015-08-04
Thanks, I updated the comment.