@andreasnoack 2015-08-04
@KristofferC Thanks for reporting this. Indeed, the fallback `chksquare` is not type stable. Furthermore, we might want to devectorize the sign determination of the determinant similarly to what you do in `det_new`. It would be great if you could open a new pr with that. (Just use `chksquare` as it should be fixed soon.)