@KristofferC 2015-08-04
Updated the benchmarks result to use #12460 instead which gives a more fair comparison.