@KristofferC 2015-08-05
Compilation of libsuitesparse failed on the mac builder?? Both travis and appveyor have been really strange lately.