@tkelman 2015-08-05
`fatal: unable to access 'https://github.com/staticfloat/homebrew-julia/': Server aborted the SSL handshake`

I restarted it, hopefully that won't keep happening.