@KristofferC 2015-08-05
@IainNZ for random matrices it should be no problem. I was thinking of problematic matrices like the Hilbert matrix.

For example:
```julia
# Create Hilbert matrix of order 4
n = 4
H = Float64[1/(i+j-1) for i=1:n, j=1:n]

# Analytical determinant
c(n) = prod([i^(n-i) for i = 1:n-1])
true_det = c(n)^n / (c(2*n))

# Compute relative error
abs((det4x4(H) - true_det)  / true_det)
# 1.104759746818661e-11
abs((det(H) - true_det) / true_det)
# 8.484729032646326e-15
```