@andreasnoack 2015-08-05
Yes. It appears that the error is much larger when computing the determinant like this. I've plotted the relative error of the determinant against the condition number
![det](https://cloud.githubusercontent.com/assets/505001/9090031/2fc7c998-3b66-11e5-8256-d43d291c9cac.png)

I think we should use formulas based on the diagonal of the U factor in the LU factorization instead. That is what I did for the "modified det3x3" line. They are slightly slower, but only slightly.

I think the easiest way to get the formulas is with SymPy. If you pass in a symbolic matrix the same function is used for calculating the determinant as for `Float64` (if you turn off pivoting).

```jl
julia> using SymPy
WARNING: could not import Base.help into PyCall

julia> A = SymPy.Sym["a$i$j" for i = 1:3, j = 1:3]
3x3 Array{SymPy.Sym,2}
⎡a₁₁  a₁₂  a₁₃⎤
⎢             ⎥
⎢a₂₁  a₂₂  a₂₃⎥
⎢             ⎥
⎣a₃₁  a₃₂  a₃₃⎦

julia> det(A)
                    ⎛      ⎛      a₁₃⋅a₂₁⎞ ⎛      a₁₂⋅a₃₁⎞          ⎞
                    ⎜      ⎜a₂₃ - ───────⎟⋅⎜a₃₂ - ───────⎟          ⎟
    ⎛      a₁₂⋅a₂₁⎞ ⎜      ⎝        a₁₁  ⎠ ⎝        a₁₁  ⎠   a₁₃⋅a₃₁⎟
a₁₁⋅⎜a₂₂ - ───────⎟⋅⎜a₃₃ - ─────────────────────────────── - ───────⎟
    ⎝        a₁₁  ⎠ ⎜                     a₁₂⋅a₂₁              a₁₁  ⎟
                    ⎜               a₂₂ - ───────                   ⎟
                    ⎝                       a₁₁                     ⎠
```