@andreasnoack 2015-08-05
Another advantage of this is that if you are a little carefull with the multiplication and division orderings, you don't use commutativity, so the check for real and complex can be avoided. Basically, the indices should "match" similarly to matrix multiplication. The second factor would be something like `A[2,2] - A[2,1]/A[1,1]*A[1,2]`.