@KristofferC 2015-08-05
That's a good idea. I will update the PR accordingly. Could you share the code you used for the condition number vs rel error plot in a gist or something? Also, what would be a good scenario to test if I got the commutativity right?