@KristofferC 2015-08-05
Question: What happens when for example `A[1,1] == 0`?