@KristofferC 2015-08-05
Just to point out, with #12460 the need for this PR was reduced significantly.