@andreasnoack 2015-08-05
The code is
```jl
Q = qr(big(randn(3,3)))[1]

vc, ve_det, ve_det3x3, ve_det3x3Alt= Float64[], Float64[], Float64[], Float64[];

for c in 10.^(5:15)
    A = map(Float64, Q*Diagonal([1;1/sqrt(c);1/c])*Q')
    push!(vc, c)
    push!(ve_det, abs(Float64(det(A)*sqrt(c)*c - 1)))
    push!(ve_det3x3, abs(Float64(Inv.det3x3(A)*sqrt(c)*c - 1)))
    push!(ve_det3x3Alt, abs(Float64(Int.det3x3_Jensen(A)*sqrt(c)*c - 1)))
end

using PyPlot
loglog(vc, [ve_det ve_det3x3 ve_det3x3Alt])
```

A way to check that the ordering is correct is to construct a matrix from its eigenvalue and vectors similarly to what I did for `A` above. Then you can check that `det(A)` is approximately the product of the eigenvalues.