@andreasnoack 2015-08-05
You are right about the problem with `A[1,1]=0` or even `A[1,1]` small.

julia> det(AA)
-0.9155469600223494 :+1: 

julia> Inv.det3x3(AA) # your verions
-0.9155469600223494 :+1: 

julia> Inv.det3x3_Jensen(AA) # my version
-0.9189646530513826 :-1: 

I should have thought about that. Pivoting is unavoidable if we want to use the LU based formula which with partial pivoting will require six branches for 3x3 and 24 branches for 4x4.

Your original version is more robust to near zero pivots, but at the cost of having larger error per condition number.